requirements.txt
torch
transformers
streamlit
dotenv
Pillow
